Introducing RawBulkAPIClient
